demo.py 加入mu_law_decode(u_law的反变换) wavenet/layers.py 修改了维度转换BUG wavenet/models.py  cost记录信息至tensorboard,模型保存和读取,移除python绘制cost函数 wavenet/utils.py mu_law_decode的实现

+++ wavenet/layers.py
@@ -20,7 +20,7 @@ def time_to_batch(inputs, rate):
     pad_left = width_pad - width
 
     perm = (1, 0, 2)
-    shape = tf.cast((width_pad / rate, -1, num_channels), tf.int32) # missing dim: batch_size * rate
+    shape = (int(width_pad / rate), -1, num_channels) # missing dim: batch_size * rate
     padded = tf.pad(inputs, [[0, 0], [pad_left, 0], [0, 0]])
     transposed = tf.transpose(padded, perm)
     reshaped = tf.reshape(transposed, shape)
